# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pkm",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/client/diagnostics": {
      "post": {
        "summary": "Client Diagnostics",
        "description": "Record a replica's self-report in the server log.\n\nA browser replica that finds its database corrupt rebuilds itself from a\nsnapshot, which destroys the evidence. Before it does, it posts what the\ndatabase said about itself (integrity checks, row counts, cursor) and\nwhich kind of client it is. The server only logs the body: the access\nlog around the line already holds the requests that led up to it.\nNothing is written to the database, so there is no journal row and no\nnudge.",
        "operationId": "client_diagnostics_api_client_diagnostics_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientDiagnosticsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Client Diagnostics Api Client Diagnostics Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/changes": {
      "get": {
        "summary": "Sync Changes",
        "operationId": "sync_changes_api_sync_changes_get",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Since"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1000,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangesPayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangesPayload": {
        "properties": {
          "reset": {
            "type": "boolean",
            "title": "Reset",
            "default": false
          },
          "generation": {
            "type": "string",
            "title": "Generation"
          },
          "plain_space_title_canonicalization": {
            "type": "boolean",
            "title": "Plain Space Title Canonicalization"
          },
          "next_since": {
            "type": "integer",
            "title": "Next Since"
          },
          "latest_seq": {
            "type": "integer",
            "title": "Latest Seq"
          },
          "pages": {
            "items": {
              "$ref": "#/components/schemas/SyncPage"
            },
            "type": "array",
            "title": "Pages"
          },
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/SyncBlock"
            },
            "type": "array",
            "title": "Blocks"
          },
          "sidebar": {
            "items": {
              "$ref": "#/components/schemas/SyncSidebarEntry"
            },
            "type": "array",
            "title": "Sidebar"
          },
          "tombstones": {
            "items": {
              "$ref": "#/components/schemas/SyncTombstone"
            },
            "type": "array",
            "title": "Tombstones"
          }
        },
        "type": "object",
        "required": [
          "generation",
          "plain_space_title_canonicalization",
          "next_since",
          "latest_seq",
          "pages",
          "blocks",
          "sidebar",
          "tombstones"
        ],
        "title": "ChangesPayload"
      },
      "ClientDiagnosticsRequest": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "error": {
            "type": "string",
            "title": "Error"
          },
          "report": {
            "additionalProperties": true,
            "type": "object",
            "title": "Report"
          },
          "client": {
            "additionalProperties": true,
            "type": "object",
            "title": "Client"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "error",
          "report",
          "client"
        ],
        "title": "ClientDiagnosticsRequest",
        "description": "A replica's self-report, sent before it rebuilds itself."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SyncBlock": {
        "properties": {
          "uid": {
            "type": "string",
            "title": "Uid"
          },
          "page_id": {
            "type": "integer",
            "title": "Page Id"
          },
          "parent_uid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Uid"
          },
          "order_idx": {
            "type": "integer",
            "title": "Order Idx"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "heading": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heading"
          },
          "view_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "numbered",
                  "document"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "View Type"
          },
          "collapsed": {
            "type": "integer",
            "title": "Collapsed"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "refs": {
            "items": {
              "$ref": "#/components/schemas/SyncRef"
            },
            "type": "array",
            "title": "Refs"
          }
        },
        "type": "object",
        "required": [
          "uid",
          "page_id",
          "parent_uid",
          "order_idx",
          "text",
          "heading",
          "view_type",
          "collapsed",
          "created_at",
          "updated_at",
          "refs"
        ],
        "title": "SyncBlock"
      },
      "SyncPage": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "created_at",
          "updated_at"
        ],
        "title": "SyncPage"
      },
      "SyncRef": {
        "properties": {
          "target_page_id": {
            "type": "integer",
            "title": "Target Page Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          }
        },
        "type": "object",
        "required": [
          "target_page_id",
          "kind"
        ],
        "title": "SyncRef"
      },
      "SyncSidebarEntry": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "order_idx": {
            "type": "integer",
            "title": "Order Idx"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "order_idx"
        ],
        "title": "SyncSidebarEntry"
      },
      "SyncTombstone": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "entity_id"
        ],
        "title": "SyncTombstone"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}